Slick-Scala-Postgres

+++ hello-slick-3.1/src/main/scala/Tables.scala
@@ -20,15 +20,15 @@ class Suppliers(tag: Tag)
 
 // A Coffees table with 5 columns: name, supplier id, price, sales, total
 class Coffees(tag: Tag)
-  extends Table[(String, Int, Double, Int, Int)](tag, "COFFEES") {
+  extends Table[(String, Int,Int, Int, Int)](tag, "COFFEES") {
 
   def name: Rep[String] = column[String]("COF_NAME", O.PrimaryKey)
   def supID: Rep[Int] = column[Int]("SUP_ID")
-  def price: Rep[Double] = column[Double]("PRICE")
+  def price: Rep[Int] = column[Int]("PRICE")
   def sales: Rep[Int] = column[Int]("SALES")
   def total: Rep[Int] = column[Int]("TOTAL")
   
-  def * : ProvenShape[(String, Int, Double, Int, Int)] =
+  def * : ProvenShape[(String, Int, Int, Int, Int)] =
     (name, supID, price, sales, total)
   
   // A reified foreign key relation that can be navigated to create a join
